refactor: dynamic data source
