--- apps/client/src/common/utils/viewUtils.ts
@@ -1,6 +1,0 @@
-export function isStringBoolean(text: string | null) {
-  if (text === null) {
-    return false;
-  }
-  return text?.toLowerCase() === 'true' || text === '1';
-}
