--- apps/client/src/features/viewers/common/viewerUtils.ts
@@ -1,25 +1,0 @@
-import { TimerType } from 'ontime-types';
-
-import type { ViewExtendedTimer } from '../../../common/models/TimeManager.type';
-
-type TimerTypeParams = Pick<ViewExtendedTimer, 'timerType' | 'current' | 'elapsed' | 'clock'>;
-
-export function getTimerByType(timerObject?: TimerTypeParams): number | null {
-  if (!timerObject) {
-    return null;
-  }
-
-  switch (timerObject.timerType) {
-    case TimerType.CountDown:
-    case TimerType.TimeToEnd:
-      return timerObject.current;
-    case TimerType.CountUp:
-      return Math.abs(timerObject.elapsed ?? 0);
-    case TimerType.Clock:
-      return timerObject.clock;
-    default: {
-      const exhaustiveCheck: never = timerObject.timerType;
-      return exhaustiveCheck;
-    }
-  }
-}
